Some aesthetic changes
A few small changes to this great program that better meet my personal ideas. I would like to remove the original symbol of the battery from the menu bar and the display of the current charge status is more suitable than the displays of the currently set limit, which is displayed via the "Quick Limits" anyway. In order to fulfill this completely, the menu item "Disable Discharging Limit" was also transferred to the "Quick Limits" as 100%. In the display "Powerflow" the use of a monospaced font is dispensed with, because it does not quite fit the design. The deviations of the text length to display the numbers arranged among each other is minimal, because the character widths of the numbers in the system font is almost "monospaced". Problematic is the "-" display that mixes up the lengths. But since + is displayed in green and - in red, the minus sign is probably dispensable. In the CLI version, the "Design Capacity" is now displayed in addition to the "Full Capacity", which makes it easy to estimate how fit the battery  still is. So these are all small changes that only affect my personal aesthetic feeling. The original program works absolutely perfectly in all areas.